FMC-188 Allow custom names for a base-map style (#252)
* FMC-188 Allow custom names for a base-map style

* FMC-188 removed key and ref
