+++ tests/components/styles.test.jsx
@@ -44,13 +44,20 @@ describe('styles', () => {
       {
         name: 'deuteranopia',
         attribution: 'Test',
-        url: 'https://test/deuteranopia/styles.json'
+        url: 'https://test/deuteranopia/styles.json',
+        displayName: 'Red-green colour deficiency'
       },
       {
         name: 'tritanopia',
         attribution: 'Test',
         url: 'https://test/tritanopia/styles.json'
+      },
+      {
+        name: 'no-display-name',
+        attribution: 'Test',
+        url: 'https://test/no-display-name/styles.json'
       }
+
     ]
   })
 
@@ -102,7 +109,7 @@ describe('styles', () => {
     fireEvent.click(container.querySelector('.fm-c-btn-more'))
 
     act(() => {
-      expect(container.querySelectorAll('.fm-c-layers__item').length).toEqual(8)
+      expect(container.querySelectorAll('.fm-c-layers__item').length).toEqual(9)
       expect(container.querySelector('.fm-c-btn-more')).toBeFalsy()
       expect(screen.getByText(/Small/)).toBeTruthy()
       expect(screen.getByText(/Medium/)).toBeTruthy()
@@ -127,6 +134,8 @@ describe('styles', () => {
     fireEvent.click(container.querySelector('.fm-c-btn-more'))
     fireEvent.click(container.querySelectorAll('.fm-c-layers__button')[0])
     fireEvent.click(container.querySelectorAll('.fm-c-layers__button')[6])
+    expect(screen.getByText(/Red-green colour deficiency/)).toBeTruthy()
+    expect(screen.getByText(/no-display-name/)).toBeTruthy()
 
     act(() => {
       expect(viewportDispatch).toHaveBeenCalledTimes(2)
